+++ packages/headless/src/controllers/commerce/product-view/headless-product-view.ts
@@ -1,6 +1,6 @@
 import {Product} from '@coveo/relay-event-types';
 import {CommerceEngine} from '../../../app/commerce-engine/commerce-engine';
-import {productView} from '../../../features/commerce/context/cart/cart-actions';
+import {productView} from '../../../features/commerce/context/view/view-actions';
 
 /**
  * The `ProductView` controller provides an interface for triggering an analytics event for a product view.
